ai-agent/requirements.txt
langchain 
gpt4all 
python-dotenv